+++ src/protocol.rs
@@ -463,6 +463,10 @@ impl ProtocolId {
         &self.0
     }
 
+    pub fn from_bytes(bytes: [u8; 32]) -> Self {
+        Self(bytes)
+    }
+
     pub fn to_hex(self) -> String {
         encode_hex(&self.0)
     }